requirements.txt
azure-ai-documentintelligence==1.0.0b2
openai
tiktoken
streamlit
semchunk